Undo extractions
