+++ app/src/main/java/com/amaze/filemanager/filesystem/operations/extract/RarExtractOperation.java
@@ -28,6 +28,7 @@ import androidx.annotation.NonNull;
 import com.amaze.filemanager.filesystem.FileUtil;
 import com.amaze.filemanager.filesystem.compressed.CompressedHelper;
 import com.amaze.filemanager.asynchronous.management.ServiceWatcherUtil;
+import com.amaze.filemanager.filesystem.operations.AbstractExtractOperation;
 import com.amaze.filemanager.utils.files.GenericCopyUtil;
 import com.github.junrar.Archive;
 import com.github.junrar.exception.RarException;
