+++ app/src/test/java/com/amaze/filemanager/filesystem/compressed/extractcontents/XzExtractOperationTest.java
@@ -1,6 +1,6 @@
 package com.amaze.filemanager.filesystem.compressed.extractcontents;
 
-import com.amaze.filemanager.filesystem.operations.extract.AbstractExtractOperation;
+import com.amaze.filemanager.filesystem.operations.AbstractExtractOperation;
 import com.amaze.filemanager.filesystem.operations.extract.XzExtractOperation;
 
 public class XzExtractOperationTest extends AbstractExtractorTest {
